[api] remove factoryreset
When OpenThread is used as a component of a product, `factoryreset` is
usually not an wanted action as there are probably other product
application related data needs to be erased. This commit removes the
`factoryreset` command and provided another `thread erase` command to
simply erase Thread settings without resetting the chip.