[6lowpan] fix forwarding to an end device
If a border router is an end device, the mesh forwarder will be destined
to an end device. In the current implementation, parents will reassemble
6lowpan packets and deliver to end device via direct transmission only.
Thus, a sleepy border router will probably miss those packets.

This commit changes the implementation so that a 6lowpan packet will be
forwarded to its final destination in 6lowpan layer, and if the
destination is a sleepy end device, the packets will be delivered via
indrect transmission.