[udp] disallow changing netif of bound sockets
The UDP socket may or may not be open from the platform layer based on
the network interface it was bound, changing the network would result in
invalid states. This commit changes the implementation of Socket::Open()
and Socket::Udp() to defer opening platform socket in bind and disallow
changing the network interface after a socket is bound to make sure the
states are consistent in OpenThread stack and the platform stack.